+++ test/test-app/src/PlayParameters.tsx
@@ -1,9 +1,0 @@
-export type PlayParameters = {
-  storyId: number;
-  apiKey: string;
-  version: number | undefined;
-  startGraphReferenceId: string;
-  charismaUrl: string;
-};
-
-export const LOCAL_STORAGE_KEY = "charisma-react-test-conversation-parameters";